requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3
flask-cors==4.0.0
pywebrtc==1.1.1
cryptography==41.0.3
python-dotenv==1.0.9
requests==2.31.0
werkzeug==2.3.7
jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
markupsafe==2.1.3
cairosvg==2.7.1
Pillow==10.0.0 